Cim

Documentation for Cim.

Cim.AbstractContour

Cim.circle

Cim.quadpts

Cim.rectangle

Cim.contr_int

Cim.contr_int_ho

Cim.get_quadpts

Cim.is_inside

Cim.show_eigenvalues!

Cim.show_quadpts!

Cim.AbstractContour — Type

contours that we plan to support: 1. ellipse 2. circle 3. rectangle (not supported yet)

Cim.circle — Type

Cim.quadpts — Type

Cim.rectangle — Type

Cim.contr_int — Method

NEP: the matrix function of the nonlinear eigenvalue problem D: the size of the NEP l: a number between k(the number of eigenvalues inside the contour) and D, k <= L <= D

Cim.contr_int_ho — Method

contour integral method with high-order moments pbar: the number of the moments (for p = 0, ..., pbar)

In this function, we follow the notations in Stefan Guttel, Francoise Tisseur, Acta Numerica, 2017 r,l : size of the probing matrices

Cim.get_quadpts — Method

get_quadpts(contour::ellipse or circle, num_quadpts::Int64)
-------------------------------------------------------
Input:
    contour    : the contour that we discretize 
    num_quadpts: the number of the quadrature nodes
Output:
    quadpts    : the struct quadpts which contains the 
                 information of the quadrature nodes

TODO: note that rectangle is different with ellipse and circle

Cim.is_inside — Method

is_inside(λ, ctr::ellipse)
is_inside(λ, ctr::circle)

Check if the eigenvalue λ is in the contour ctr to avoid spurious eigenvalues.

Cim.show_eigenvalues! — Method

show_eigenvalues(ax, eigvals::AbstractArray)

Plot the eigenvalues on complex plane

Cim.show_quadpts! — Method

show_quadpts!(ax, pts::quadpts)

Plot the quadrature points pts