API

Cim.AbstractContour

Cim.Qep

Cim.Qep

Cim.Qep

Cim.quadpts

Cim.cim

Cim.cim

Cim.contr_int_ho

Cim.get_quadpts

Cim.get_quadpts

Cim.get_quadpts

Cim.is_inside

Cim.show_contour!

Cim.show_contour!

Cim.show_eigenvalues!

Cim.show_quadpts!

Cim.AbstractContour — Type

contours that we plan to support: 1. ellipse 2. circle 3. rectangle (not supported yet)

Cim.Qep — Type

Qep{T}

Construct quadratic eigenvlaue problems

$(\mathbf{A}_{0} + \alpha \mathbf{A}_{1} + \alpha^2 \mathbf{A}_{2}) \mathbf{u} = 0.$

Fields

A₀: the matrix in the zero order term

A₁: the matrix in the first order term

A₂: the matrix in the second order term

Cim.Qep — Method

(q::Qep{T})(α::S) where {T, S}

Compute the matrix of the quadratic eigenvalue problem at specific value α.

Cim.Qep — Method

Qep{T}(; 
A₀::SparseMatrixCSC{T, Int}, 
A₁::SparseMatrixCSC{T, Int}, 
A₂::SparseMatrixCSC{T, Int}) where T

Use keywords arguments to construct qaudratic eigenvalue problems.

Cim.quadpts — Type

quadpts

Quadrature points of the trapezoidal rule on the contours.

Properties

N: the number of the quadrature nodes

nodes: quadrature nodes size of N x 2

nodes_prime: derivative of the parametrization size of N x 2

Cim.cim — Method

cim(ctr::AbstractContour, nep::Function, D, l::Int64; n=50, tol=1e-12)

Contour integral method to calculate the eigenvalues inside the contour ctr

Arugments

ctr: the contour

nep: the nonlinear eigenvalue problem

D: the scale of the nep

l: the number of columns of random matrix

n: the number of the quadrature points (here we use the trapezoid rule)

tol: tolerance to determine the number of nonzero singular values

Reference

Wolf-Jurgen Beyn, An integral method for solving NEPs, Linear Algebra Appl., 2012.

Cim.cim — Method

cim(ctr::AbstractContour, nep::Qep, d::Int64, l::Int64; n=50, tol=1e-12)

Use the contour integral method to solve quadratic eigenvalue problems.

Cim.contr_int_ho — Method

contr_int_ho()

Contour integral method with high-order moments

Arguments

In this function, we follow the notations in Stefan Guttel, Francoise Tisseur, Acta Numerica, 2017

ctr:

nep:

D:

r, l: size of the probing matrices

pbar: the number of the moments (for p = 0, ..., pbar)

Reference

Stefan Guttel, Francoise Tisseur, The NEP, Acta Numerica, 2017.

Cim.get_quadpts — Method

get_quadpts(ctr::circle, num_quadpts::Int64)

Get the quadrature points on the circle ctr. Here we use trapezoidal rule.

Arguments

ctr: the contour that we discretize

num_quadpts: the number of the quadrature nodes

TODO: note that rectangle is different with ellipse and circle

Cim.get_quadpts — Method

get_quadpts(ctr::ellipse, num_quadpts::Int64)

Get the quadrature points on the ellipse ctr. Here we use trapezoidal rule.

Arguments

ctr`: the contour that we discretize

num_quadpts: the number of the quadrature nodes

TODO: note that rectangle is different with ellipse and circle

Cim.is_inside — Method

is_inside(λ, ctr::ellipse)
is_inside(λ, ctr::circle)

Check if the eigenvalue λ is in the contour ctr to avoid spurious eigenvalues.

Cim.show_contour! — Method

show_contour!(ax, ctr::circle)

Plot the contour ctr::circle on Axis ax by using CairoMakie.

Cim.show_contour! — Method

show_contour!(ax, ctr::ellipse)

Plot the contour ctr::ellipse on Axis ax by using CairoMakie.

Cim.show_eigenvalues! — Method

show_eigenvalues(ax, eigvals::AbstractArray)

Plot the eigenvalues on complex plane on Axis ax by using CairoMakie.

Cim.show_quadpts! — Method

show_quadpts!(ax, pts::quadpts)

Plot the quadrature points pts on Axis ax by using CairoMakie.